backend/requirements.txt
asgiref==3.8.1
click==8.1.7
deptry==0.16.1
Django==5.0.6
django-cors-headers==4.3.1
django-dotenv==1.4.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
packaging==24.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.1
